requirements.txt
opencv-python-headless
numpy
pydantic
python-dotenv
azure-ai-documentintelligence
google-genai
pdf2image
Pillow
